ajout de tests + fix bug rechercherBars + ajout de ToString + fin protocole egalité + encapsulation dictionnary + dataTemplate sur nom des villes dans comboBox

+++ SearchBars/MetierSearchBars/IBoisson.cs
@@ -11,6 +11,6 @@ namespace MetierSearchBars
         TypeBoisson Type { get; }
         string Nom { get; }
         double Prix { get; }
-        int DegreAlcool { get; }
+        double DegreAlcool { get; }
     }
 }
